test: restructure tests and increase coverage
- Move `src/player/tests.rs` into `src/player/tests/` subdirectory and split into `player_tests.rs`, `builder_tests.rs`, and `convenience_tests.rs`.
- Split `src/group/tests/mod.rs` into `manager_tests.rs` and `device_group_tests.rs` and add tests for edge cases like getting or deleting non-existent groups.
- Expand `src/client/tests/unified_tests.rs` to cover playback methods and disconnected state validation.
- All test restructuring follows the documentation requirements for keeping files manageable and enforcing API specs.
