+++ src/player/tests/mod.rs
@@ -1,0 +1,3 @@
+mod builder_tests;
+mod convenience_tests;
+mod player_tests;
